fix hardware ID token
The HW token is supposed to prevent the hardware engine from being
re-used across multiple instances of the hash object.

The original implementation allowed some re-use leakge in the case
that the hash object was cloned. Furthermore, it did not release
the engine if it was acquired but not used prior to the object going
out of scope.

This patch fixes these issues.